+++ PublicApi.DTO.V1/DailyNutritionIntake.cs
@@ -10,14 +10,14 @@ namespace PublicApi.DTO.V1
     public class DailyNutritionIntakeCreate
     {
         [Range(1, 10000)]
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         [Range(1, 1000)]
-        public int Protein { get; set; }
+        public float Protein { get; set; }
         [Range(1, 1000)]
-        public int Fats { get; set; }
+        public float Fats { get; set; }
         
         [Range(1, 1000)]
-        public int Carbohydrates { get; set; }
+        public float Carbohydrates { get; set; }
     }
 
     public class DailyNutritionIntakeEdit : DailyNutritionIntakeCreate
